without_oauth/requirements.txt
flask
google-cloud-datastore
gunicorn